# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Orvix Enterprise Mail API
  version: "1.0.0"
  description: |
    SaaS API for Orvix Enterprise Mail.
    All API endpoints are under /api/v1.
    Authentication uses JWT Bearer tokens or API keys.
servers:
  - url: https://api.orvix.email/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: api_key
      description: API key with 'orv_' prefix
...
  schemas:
...
    AuditEntry:
      type: object
      description: Rich extended audit entry (orvix_audit row)
      required: [id, actor, action, result, timestamp]
      properties:
        id:
          type: integer
          format: int64
        actor:
          type: string
          example: "user:42"
        actor_id:
          type: integer
        actor_role:
          type: string
        tenant_id:
          type: integer
        action:
          type: string
          example: organization.create
        target:
          type: string
        target_id:
          type: integer
        result:
          type: string
          enum: [success, failure]
        reason:
          type: string
        before:
          type: string
        after:
          type: string
        request_id:
          type: string
        ip:
          type: string
        user_agent:
          type: string
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /audit/logs/{id}:
    get:
      summary: One extended audit entry
      operationId: getAuditEntry
      tags: [Security]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Extended audit entry (same contract as list rows)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEntry'
        '400':
          description: Invalid id (code VALIDATION_FAILED)
        '401':
          description: Authentication required
        '403':
          description: Platform role required
        '404':
          description: Entry not found (code NOT_FOUND)